Migrate natlab from tcli to UniFFI bindings
So far, natlab tests have been using tcli to interact with libtelio.
Now that we have FFI bindings that can be easily generated from source
as part of the build, it makes sense to migrate natlab to UniFFI
bindings, as a form of dogfooding.

The main thing that is new here is the introduction of Pyro5 to enable
communication between the host and the different containers in which
the tests are running, since in general, libtelio is running in some
container(s) and/or VM(s) whereas the test is run from the host and
there needs to be some way for the host the control what libtelio is
doing. The bindings we are using are in python, so some kind of RPC or
remote object framework is needed. Pyro5 is one such framework that is
easy and low friction. In the containers and VMs, libtelio_remote.py
is running, which in turn is running an instance of libtelio through
generated bindings. Then on the host side, libtelio_proxy.py is set
up to communicate with libtelio_remote.py through standard Pyro5
mechanisms. As a result, we can add a property to our test Client
that has an instance of a proxy, allowing us to easily communicate
with the remote as if it was also running on the host.

Since a test Client already abstracts away the communication with tcli,
doing the migration doesn't involve that many code changes inside the
actual tests.