Change underlying type of keys in FFI
Due to a bug in the C# generator for UniFFI, we cannot use the UniFFI
type `bytes` in type aliases. As a workaround, we change the underlying
type to `sequence<u8>`, which should be the same under-the-hood, just
a flat list of bytes, but without the syntactic sugar and potential
optimizations that languages might add for their specific byte array
types. The intention is that this fix will be temporary until the
C# generator has been fixed to emit working code in this scenario